Discover nodes run by systemd and tail journald logs (#114)
* Add systemd+journal capabilities in node discovery

Refactors the Chain interface to remove dependencies specific to how the
node process is managed (i.e. docker, systemd). Adds a mechanism to
detect how the node is set up in the host and lazily starts the required
watchers.

The discovery mechanism prioritizes docker if it detects a dockerized
node that is managed by systemd. The agent will keep retrying the
discovery process until the node appears in systemd or docker.