Fix bug causing agent to skip node version configuration when node is discovered (#139)
* Fix bug causing agent to skip node version configuration when node is discovered

Fixes a bug where agent would skip node version configuration when
network and node role was previously set during node discovery. This
implements a util function used to extract the node version from a PEF
endpoint using a regex. The latter is used as a default, but if the
regex doesn't match, the version is extracted from the container's image
tag for backwards compatibility. `ReconfigurationByDocker/Systemd()` are now
ignoring network and node role presence and will always run a full
configuration when called.